Moved package_testcase to a method of ProfilingCase
